split into detail repository
