+++ src/packages/user/user/repository/detail/index.ts
@@ -1,0 +1,1 @@
+export { UmbUserDetailRepository } from './user-detail.repository.js';
