+++ src/packages/user/user/repository/index.ts
@@ -5,3 +5,4 @@ export * from './unlock/unlock-user.repository.js';
 export * from './invite/invite-user.repository.js';
 export * from './user.repository.js';
 export * from './item/index.js';
+export * from './detail/index.js';
